requirements.txt
fastapi==0.95.2
jinja2==3.1.2
python-dotenv==1.0.0
uvicorn==0.22.0
